backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
certifi==2025.10.5
cffi==2.0.0
click==8.2.1
cryptography==46.0.1
ecdsa==0.19.1
fastapi==0.116.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
passlib==1.7.4
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
